ac3

+++ src/main/java/org/apache/ibatis/cache/decorators/ScheduledCache.java
@@ -32,7 +32,7 @@ public class ScheduledCache implements Cache {
 
   public ScheduledCache(Cache delegate) {
     this.delegate = delegate;
-    this.clearInterval = 60 * 60 * 1000; // 1 hour
+    this.clearInterval = (long) 60 * 60 * 1000; // 1 hour
     this.lastClear = System.currentTimeMillis();
     this.sDelegate = new SerializedCache(delegate);
   }
